+++ CHANGELOG.md
@@ -11,6 +11,11 @@ Work-in-progress :thunder:
 + **mail** - added `updateEmailFlags`.
 + **madvoc** - added flag to `RestActionNamingStrategy`.
 + **petite** - added `PetiteContainer#invokeMethod`.
++ **petite** - added cache for external types that are not registered into the container.
+
+### Breaking changes
+
++ **core** - `TypeCache` made better. `put()` removed in favor of `get()`.
 
 
 ## [4.3.2](https://github.com/oblac/jodd/compare/v4.3.1...4.3.2)